+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3 (11.July.2017)
+
+- Fixed [bug #1](https://github.com/KnowTheCode/UpGulp/issues/1) to disable `cleanStyles()` when the package is not theme.
+- Passing in `settings` to each function in `gulp/tasks/styles.js` for consistency.
+
+## 1.0.2
+
+- DocBlock cleanup
+
 ## 1.0.1
 
 - distFilename configuration added